Update project Race

+++ BasicProgramming/11-GAME/Race/RaceObject/Boots.h
@@ -6,9 +6,9 @@
 class Boots: public Transport
 { 
 private:
-	int DrivingTime;	// Время движения до отдыха
-	int RestTime1;		// Длительность отдыха первый раз
-	int RestTime2;		// Длительность отдыха все последующие разы
+	double DrivingTime;		// Время движения до отдыха
+	double RestTime1;		// Длительность отдыха первый раз
+	double RestTime2;		// Длительность отдыха все последующие разы
 public:
 	Boots();
