+++ BasicProgramming/11-GAME/Race/RaceObject/Camel.h
@@ -1,12 +1,13 @@
 #pragma once
 #include "Transport.h"
 
+// Верблюд
 class Camel : public Transport
 {
 private:
-	int DrivingTime;	// Время движения до отдыха
-	int RestTime1;		// Длительность отдыха первый раз
-	int RestTime2;		// Длительность отдыха все последующие разы
+	double DrivingTime;		// Время движения до отдыха
+	double RestTime1;		// Длительность отдыха первый раз
+	double RestTime2;		// Длительность отдыха все последующие разы
 public:
 	Camel();
