+++ BasicProgramming/11-GAME/Race/RaceObject/SpeedCamel.h
@@ -5,10 +5,10 @@
 class SpeedCamel :  public Transport
 {
 private:
-	int DrivingTime;	// Время движения до отдыха
-	int RestTime1;		// Длительность отдыха первый раз
-	int RestTime2;		// Длительность отдыха во второй раз
-	int RestTime3;		// Длительность отдыха все последующие разы
+	double DrivingTime;	// Время движения до отдыха
+	double RestTime1;		// Длительность отдыха первый раз
+	double RestTime2;		// Длительность отдыха во второй раз
+	double RestTime3;		// Длительность отдыха все последующие разы
 public:
 	SpeedCamel();
